Extract material and geometry creation
