+++ .gitignore
@@ -1,3 +1,4 @@
 .DS_Store
 **/.DS_Store
 **/**/.DS_Store
+.idea
